feat: python websocket_gamium_service support
