+++ client/python/gamium/gamium/__init__.py
@@ -1,5 +1,7 @@
 from gamium.gamium_client import *
-from gamium.gamium_service import *
+from gamium.igamium_service import *
+from gamium.tcp_gamium_service import *
+from gamium.websocket_gamium_service import *
 from gamium.locator.by import *
 from gamium.locator.locator import *
 from gamium.locator.rpc_by import *
