+++ client/python/gamium/pyproject.toml
@@ -9,6 +9,7 @@ readme = "README.md"
 python = "^3.9"
 flatbuffers = "22.10.26"
 numpy = "^1.24.0"
+websockets = "^11.0.3"
 
 
 [build-system]
